# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DEPOT API",
        "contact": {
            "name": "oussema khlifi",
            "email": "khoussemaa@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/invoices/create": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Invoice store",
                "operationId": "dc2751d6fb8093996c56026541116ca6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "provider_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "recivedat": {
                                        "type": "date"
                                    },
                                    "price": {
                                        "type": "decimal"
                                    },
                                    "discount": {
                                        "type": "decimal"
                                    },
                                    "total": {
                                        "type": "decimal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/invoices/invoice/{id}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Invoice show",
                "operationId": "26e17861dded81979a1dc84c847d34ae",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/invoices/update/{id}": {
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Invoice update",
                "operationId": "4c34e85e186246cac33ad55ad23d42ba",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "integer"
                                    },
                                    "provider_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    },
                                    "recivedat": {
                                        "type": "date"
                                    },
                                    "price": {
                                        "type": "decimal"
                                    },
                                    "discount": {
                                        "type": "decimal"
                                    },
                                    "total": {
                                        "type": "decimal"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/invoices/delete/{id}": {
            "delete": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Invoice delete",
                "operationId": "0333d40fe2c17f5a6959e8e574e01576",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal server error"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InvoiceSchema": {
                "title": "Invoice Model",
                "description": "Invoice Model",
                "properties": {
                    "product_id": {
                        "description": "product id",
                        "type": "integer",
                        "example": "1"
                    },
                    "provider_id": {
                        "description": "provider id",
                        "type": "integer",
                        "example": "1"
                    },
                    "status": {
                        "description": "invoice status",
                        "type": "string",
                        "example": "pending"
                    },
                    "quantity": {
                        "description": "product quantity",
                        "type": "integer",
                        "example": "10"
                    },
                    "recivedat": {
                        "description": "product delivery date",
                        "type": "date",
                        "example": "2021/08/11 11:045:20"
                    },
                    "price": {
                        "description": "product price",
                        "type": "decimal",
                        "example": "455.2"
                    },
                    "discount": {
                        "description": "product discount",
                        "type": "decimal",
                        "example": "1500"
                    },
                    "total": {
                        "description": "invoice total",
                        "type": "decimal",
                        "example": "1500.57"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "query",
                "description": "depot id",
                "schema": {
                    "type": "number"
                }
            }
        }
    }
}